Services/AppraisalGuide/pyproject.toml
[project]
name = "appraisalguide"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.4",
    "langchain>=1.1.3",
    "langchain-google-genai>=4.0.0",
    "langgraph>=1.0.5",
    "openpyxl>=3.1.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sse-starlette>=3.0.4",
    "uvicorn>=0.38.0",
]
